+++ src/components/CreateProposal/CreateProposal.scss.d.ts
@@ -5,8 +5,7 @@ export const error: string;
 export const contributionReward: string;
 export const reward: string;
 export const externalTokenSelect: string;
-export const contributionType: string;
-export const contributionTypeDropdown: string;
+export const permissionCheckbox: string;
 export const requiredMarker: string;
 export const infoTooltip: string;
 export const errorMessage: string;
@@ -18,8 +17,21 @@ export const tokenAmount: string;
 export const userAvatar: string;
 export const transferIcon: string;
 export const transferTotals: string;
+export const schemeFields: string;
+export const schemeSelectContainer: string;
 export const exitProposalCreation: string;
 export const submitProposal: string;
 export const loading: string;
 export const sendIcon: string;
 export const createProposalActions: string;
+export const schemeRegistrar: string;
+export const schemeRegistrarSidebar: string;
+export const schemeItem: string;
+export const schemeTitle: string;
+export const addScheme: string;
+export const selected: string;
+export const schemeRegistrarForm: string;
+export const otherScheme: string;
+export const schemeSelect: string;
+export const permissions: string;
+export const permissionsLabel: string;
